+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialNetDriver.h
@@ -206,7 +206,7 @@ private:
 
 	void InitiateConnectionToSpatialOS(const FURL& URL);
 
-	bool ShouldOverrideReceptionistHost(USpatialGameInstance* gameInstance, FString host);
+	bool ShouldOverrideReceptionistHost(USpatialGameInstance* GameInstance, FString Host);
 
 	void InitializeSpatialOutputDevice();
 	void CreateAndInitializeCoreClasses();