+++ SpatialGDK/Source/SpatialGDK/Public/Interop/Connection/SpatialWorkerConnection.h
@@ -20,7 +20,7 @@ DECLARE_LOG_CATEGORY_EXTERN(LogSpatialWorkerConnection, Log, All);
 class USpatialGameInstance;
 class UWorld;
 
-enum class SpatialConnectionType
+enum class ESpatialConnectionType
 {
 	Receptionist,
 	LegacyLocator,
@@ -61,7 +61,7 @@ public:
 	FString GetWorkerId() const;
 	const TArray<FString>& GetWorkerAttributes() const;
 
-	void SetConnectionType(SpatialConnectionType newType);
+	void SetConnectionType(ESpatialConnectionType newType);
 
 	FReceptionistConfig ReceptionistConfig;
 	FLocatorConfig LocatorConfig;
@@ -75,7 +75,7 @@ private:
 	void OnPreConnectionFailure(const FString& Reason);
 	void OnConnectionFailure();
 
-	SpatialConnectionType GetConnectionType() const;
+	ESpatialConnectionType GetConnectionType() const;
 
 	void CacheWorkerAttributes();
 
@@ -118,5 +118,5 @@ private:
 	// RequestIds per worker connection start at 0 and incrementally go up each command sent.
 	Worker_RequestId NextRequestId = 0;
 
-	SpatialConnectionType ConnectionType = SpatialConnectionType::Receptionist;
+	ESpatialConnectionType ConnectionType = ESpatialConnectionType::Receptionist;
 };