+++ CHANGELOG.md
@@ -5,7 +5,6 @@ The format of this Changelog is based on [Keep a Changelog](https://keepachangel
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased-`x.y.z`] - 2019-xx-xx
-- "receptionistHost <URL>" can now be used as a command line arg to change the initial connection URL for a PIE client.
 
 ### Features:
 - Added an AuthorityIntent component to be used in the future for UnrealGDK code to control loadbalancing.
@@ -67,6 +66,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Muticast RPCs that are sent shortly after an actor is created are now correctly processed by all clients.
 - When replicating an actor, the owner's Spatial position will no longer be used if it isn't replicated.
 - Fixed a crash upon checking out an actor with a deleted static subobject.
+- "receptionistHost <URL>" command line arg now correctly overwrites initial connection URL for PIE clients and will not affect other client types.
+- The receptionist will now be used for appropriate URLs after connecting to a locator URL.
 
 ## [`0.6.2`] - 2019-10-10